Docs and NPM build fix
