+++ package.json
@@ -21,6 +21,7 @@
     "remark": "^15.0.1",
     "remark-html": "^16.0.1",
     "sanitize-html": "^2.17.0",
+    "server-only": "^0.0.1",
     "zod": "^3.25.76"
   },
   "devDependencies": {
